requirements.txt
djangorestframework>= "*"
Django>="*"
psycopg2-binary>=2.8
appdirs==1.4.4
certifi==2020.12.5
distlib==0.3.1
filelock==3.0.12
pipenv==2020.11.15
psycopg2==2.8.6
six==1.15.0
virtualenv==20.2.2
virtualenv-clone==0.5.4
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.5
djangorestframework==3.12.2
flake8==3.8.4
idna==2.10
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.20.0
pandas==1.2.1
psycopg2==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
six==1.15.0
sodapy==2.1.0
sqlparse==0.4.1
urllib3==1.26.3
